examples/requirements.txt
python-dotenv
pynacl
requests
